+++ stwo_cairo_prover/crates/prover/src/cairo_air/mod.rs
@@ -1,3 +1,7 @@
+use serde::{Deserialize, Serialize};
+
+use crate::components::prelude::PreProcessedTrace;
+
 pub(crate) mod air;
 pub(crate) mod blake;
 pub(crate) mod builtins_air;
@@ -7,7 +11,6 @@ pub(crate) mod pedersen;
 pub(crate) mod poseidon;
 pub(crate) mod preprocessed;
 pub mod preprocessed_utils;
-pub mod prover;
 pub mod range_checks_air;
 pub(crate) mod relations;
 
@@ -15,3 +18,22 @@ pub(crate) mod relations;
 pub mod verifier;
 
 pub use air::CairoProof;
+
+/// The preprocessed trace used for the prover.
+// TODO(Ohad): move somewhere else.
+#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub enum PreProcessedTraceVariant {
+    Canonical,
+    CanonicalWithoutPedersen,
+}
+impl PreProcessedTraceVariant {
+    pub fn to_preprocessed_trace(&self) -> PreProcessedTrace {
+        match self {
+            PreProcessedTraceVariant::Canonical => PreProcessedTrace::canonical(),
+            PreProcessedTraceVariant::CanonicalWithoutPedersen => {
+                PreProcessedTrace::canonical_without_pedersen()
+            }
+        }
+    }
+}
